franchise_mas/pyproject.toml
[project]
name = "franchise-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]
